Extend Kestrel tests and move them to their own file
