+++ tests/Aspire.Hosting.Tests/ProjectResourceTests.cs
@@ -362,148 +362,6 @@ public class ProjectResourceTests
         Assert.False(config.ContainsKey("ASPNETCORE_HTTPS_PORT"));
     }
 
-    private sealed class ProjectWithProfileEndpointAndKestrelHttpEndpoint : BaseProjectWithProfileAndConfig
-    {
-        public ProjectWithProfileEndpointAndKestrelHttpEndpoint()
-        {
-            Profiles = new()
-            {
-                ["OnlyHttp"] = new LaunchProfile
-                {
-                    ApplicationUrl = "http://localhost:5031",
-                }
-            };
-            JsonConfigString = """
-            {
-              "Kestrel": {
-                "Endpoints": {
-                  "SomeHttpEndpoint": { "Url": "http://*:5002" }
-                }
-              }
-            }
-            """;
-        }
-    }
-
-    [Fact]
-    public async Task SingleKestrelHttpEndpointIsNamedHttpAndOverridesProfile()
-    {
-        var appBuilder = CreateBuilder(operation: DistributedApplicationOperation.Run);
-        var projectBuilder = appBuilder.AddProject<ProjectWithProfileEndpointAndKestrelHttpEndpoint>("projectName");
-        using var app = appBuilder.Build();
-        var appModel = app.Services.GetRequiredService<DistributedApplicationModel>();
-        var projectResources = appModel.GetProjectResources();
-        var resource = Assert.Single(projectResources);
-
-        Assert.Collection(
-            resource.Annotations.OfType<EndpointAnnotation>(),
-            a =>
-            {
-                // Endpoint is named "http", because there is only one Kestrel http endpoint
-                Assert.Equal("http", a.Name);
-                Assert.Equal("http", a.UriScheme);
-                Assert.Equal(5002, a.Port);
-            }
-            );
-
-        var config = await EnvironmentVariableEvaluator.GetEnvironmentVariablesAsync(resource);
-
-        // The Kestrel endpoint overrides the profile endpoint
-        Assert.Equal("http://localhost:5002", config["ASPNETCORE_URLS"]);
-    }
-    private sealed class ProjectWithKestrelHttpsEndpoint : BaseProjectWithProfileAndConfig
-    {
-        public ProjectWithKestrelHttpsEndpoint()
-        {
-            JsonConfigString = """
-            {
-              "Kestrel": {
-                "Endpoints": {
-                  "SomeHttpsEndpoint": { "Url": "https://*:7002" }
-                }
-              }
-            }
-            """;
-        }
-    }
-
-    [Fact]
-    public async Task SingleKestrelHttpsEndpointIsNamedHttps()
-    {
-        var appBuilder = CreateBuilder(operation: DistributedApplicationOperation.Run);
-        var projectBuilder = appBuilder.AddProject<ProjectWithKestrelHttpsEndpoint>("projectName");
-        using var app = appBuilder.Build();
-        var appModel = app.Services.GetRequiredService<DistributedApplicationModel>();
-        var projectResources = appModel.GetProjectResources();
-        var resource = Assert.Single(projectResources);
-
-        Assert.Collection(
-            resource.Annotations.OfType<EndpointAnnotation>(),
-            a =>
-            {
-                // Endpoint is named "https"", because there is only one Kestrel https endpoint
-                Assert.Equal("https", a.Name);
-                Assert.Equal("https", a.UriScheme);
-                Assert.Equal(7002, a.Port);
-            }
-            );
-
-        var config = await EnvironmentVariableEvaluator.GetEnvironmentVariablesAsync(resource);
-
-        Assert.Equal("https://localhost:7002", config["ASPNETCORE_URLS"]);
-    }
-    private sealed class ProjectWithMultipleHttpKestrelEndpoints : BaseProjectWithProfileAndConfig
-    {
-        public ProjectWithMultipleHttpKestrelEndpoints()
-        {
-            JsonConfigString = """
-            {
-              "Kestrel": {
-                "EndpointDefaults": {
-                  "Protocols": "Http2"
-                },
-                "Endpoints": {
-                  "FirstHttpEndpoint": { "Url": "http://*:5002" },
-                  "SecondHttpEndpoint": { "Url": "http://*:5003" }
-                }
-              }
-            }
-            """;
-        }
-    }
-
-    [Fact]
-    public async Task MultipleKestrelHttpEndpointsKeepTheirNames()
-    {
-        var appBuilder = CreateBuilder(operation: DistributedApplicationOperation.Run);
-        var projectBuilder = appBuilder.AddProject<ProjectWithMultipleHttpKestrelEndpoints>("projectName");
-        using var app = appBuilder.Build();
-        var appModel = app.Services.GetRequiredService<DistributedApplicationModel>();
-        var projectResources = appModel.GetProjectResources();
-        var resource = Assert.Single(projectResources);
-
-        Assert.Collection(
-            resource.Annotations.OfType<EndpointAnnotation>(),
-            a =>
-            {
-                // Endpoints keep their config names because there are multiple Kestrel http endpoints
-                Assert.Equal("FirstHttpEndpoint", a.Name);
-                Assert.Equal("http", a.UriScheme);
-                Assert.Equal(5002, a.Port);
-            },
-            a =>
-            {
-                Assert.Equal("SecondHttpEndpoint", a.Name);
-                Assert.Equal("http", a.UriScheme);
-                Assert.Equal(5003, a.Port);
-            }
-            );
-
-        var config = await EnvironmentVariableEvaluator.GetEnvironmentVariablesAsync(resource);
-
-        Assert.Equal("http://localhost:5002;http://localhost:5003", config["ASPNETCORE_URLS"]);
-    }
-
     [Fact]
     public void DisabledForwardedHeadersAddsAnnotationToProject()
     {
@@ -653,7 +511,7 @@ public class ProjectResourceTests
             arg => Assert.Equal("http://localhost:1234", arg));
     }
 
-    private static IDistributedApplicationBuilder CreateBuilder(string[]? args = null, DistributedApplicationOperation operation = DistributedApplicationOperation.Publish)
+    internal static IDistributedApplicationBuilder CreateBuilder(string[]? args = null, DistributedApplicationOperation operation = DistributedApplicationOperation.Publish)
     {
         var resolvedArgs = new List<string>();
         if (args != null)
@@ -678,7 +536,7 @@ public class ProjectResourceTests
         public LaunchSettings? LaunchSettings { get; set; }
     }
 
-    private abstract class BaseProjectWithProfileAndConfig : IProjectMetadata
+    internal abstract class BaseProjectWithProfileAndConfig : IProjectMetadata
     {
         protected Dictionary<string, LaunchProfile>? Profiles { get; set; } = new();
         protected string? JsonConfigString { get; set; }
